requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
Flask==3.1.0
future @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
googleapis-common-protos==1.69.2
grpcio==1.71.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
macholib @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.2
objprint==0.3.0
opentelemetry-api==1.31.1
opentelemetry-distro==0.52b1
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asyncio==0.52b1
opentelemetry-instrumentation-click==0.52b1
opentelemetry-instrumentation-dbapi==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-grpc==0.52b1
opentelemetry-instrumentation-jinja2==0.52b1
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sqlite3==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-urllib==0.52b1
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-wsgi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
otlp-package @ file:///Users/sravani/Documents/py_pt/otlp_package
packaging==24.2
protobuf==5.29.4
python-otlp @ file:///Users/sravani/Documents/python_otlp
requests==2.32.3
six @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
structlog==25.2.0
thrift==0.21.0
typing_extensions==4.13.0
urllib3==1.26.16
viztracer==0.15.6
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
